+++ WorkingApp/Dockerfile
@@ -4,16 +4,16 @@ EXPOSE 80
 
 FROM microsoft/dotnet:2.1-sdk AS build
 WORKDIR /src
-COPY WebApplication4/WebApplication4.csproj WebApplication4/
-RUN dotnet restore WebApplication4/WebApplication4.csproj
+COPY WorkingApp/WorkingApp.csproj WorkingApp/
+RUN dotnet restore WorkingApp/WorkingApp.csproj
 COPY . .
-WORKDIR /src/WebApplication4
-RUN dotnet build WebApplication4.csproj -c Release -o /app
+WORKDIR /src/WorkingApp
+RUN dotnet build WorkingApp.csproj -c Release -o /app
 
 FROM build AS publish
-RUN dotnet publish WebApplication4.csproj -c Release -o /app
+RUN dotnet publish WorkingApp.csproj -c Release -o /app
 
 FROM base AS final
 WORKDIR /app
 COPY --from=publish /app .
-ENTRYPOINT ["dotnet", "WebApplication4.dll"]
+ENTRYPOINT ["dotnet", "WorkingApp.dll"]
