+++ WorkingApp/Properties/launchSettings.json
@@ -17,7 +17,7 @@
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
-    "WebApplication4": {
+    "WorkingApp": {
       "commandName": "Project",
       "launchBrowser": true,
       "launchUrl": "api/values",
