+++ docker-compose.override.yml
@@ -1,7 +1,7 @@
 version: '3.4'
 
 services:
-  webapplication4:
+  workingapp:
     environment:
       - ASPNETCORE_ENVIRONMENT=Development
     ports:
